Stop reading/writing state/node files.  Fixes #5.
CVE-2013-4184 discusses the insecure usage of `/tmp/.UUID_STATE` and
`/tmp/.UUID_NODEID`, due to a potential symlink attack.

Unfortunately, this code is old enough that none of the current
maintainers fully understand the history of the code, nor do we have
sufficient context from the commit history.

Rather than attempting to address this vulnerability by changing the
location of where these files are written, or by adding additional logic
to ensure that they are handled securely, this patch addresses the issue
by eliminating the need for these two files completely.